add town areas uploader

+++ app/models/town_counter.rb
@@ -5,5 +5,5 @@ class TownCounter
 
   field :citizens, type: Integer
   field :house_holdings, type: Integer
-  field :square, type: Integer
+  field :square, type: Float
 end
